+++ CHANGELOG.md
@@ -5,6 +5,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+- The JSON response body will only be documented when the content type is `application/json` [#11] ([@2potatocakes])
 
 ## [0.14.0] - 2018-07-21